move tsconfig from typescript to javascript
