+++ src/javascript/index.ts
@@ -5,3 +5,4 @@ export * from './node-project';
 export * from './npm-config';
 export * from './upgrade-dependencies';
 export * from './projenrc';
+export * from './typescript-config';
