+++ src/typescript/index.ts
@@ -1,4 +1,3 @@
 export * from './projenrc';
 export * from './typescript';
-export * from './typescript-config';
 export * from './typescript-typedoc';
