+++ src/web/react.ts
@@ -2,10 +2,10 @@ import * as path from 'path';
 import { Component } from '../component';
 import { FileBase, FileBaseOptions, IResolver } from '../file';
 import { NodeProject, NodeProjectOptions } from '../javascript';
+import { TypeScriptJsxMode, TypeScriptModuleResolution } from '../javascript/typescript-config';
 import { SampleDir } from '../sample-file';
 import { SourceCode } from '../source-code';
 import { TypeScriptAppProject, TypeScriptProjectOptions } from '../typescript';
-import { TypeScriptJsxMode, TypeScriptModuleResolution } from '../typescript/typescript-config';
 import { deepMerge } from '../util';
 
 export interface ReactRewireOptions {
